Consolidate saved state transforms into ProfileStateRegistry
