+++ src/platform/plugins/shared/discover/common/context_awareness/create_profile_state_registry.ts
@@ -10,6 +10,7 @@
 import { ProfileStateRegistry } from './profile_state';
 import { EXAMPLE_PROFILE_STATE_DEF } from './profile_state_definitions/example_profile_state';
 import { METRICS_STATE_DEF } from './profile_state_definitions/metrics_grid_profile_state';
+import { METRICS_GRID_SAVED_STATE_TRANSFORM } from './profile_saved_state_transforms/metrics_grid_saved_state_transform';
 
 /** Creates the Discover profile state registry with all supported definitions registered. */
 export const createProfileStateRegistry = () => {
@@ -17,6 +18,7 @@ export const createProfileStateRegistry = () => {
 
   registry.registerDefinition(EXAMPLE_PROFILE_STATE_DEF);
   registry.registerDefinition(METRICS_STATE_DEF);
+  registry.registerTransform(METRICS_GRID_SAVED_STATE_TRANSFORM);
 
   return registry;
 };
