+++ src/platform/plugins/shared/discover/common/context_awareness/index.ts
@@ -15,6 +15,9 @@ export {
   type ProfileStateDefinition,
   type ProfileStateDescriptor,
   type ProfileStateMap,
+  createProfileSavedStateTransform,
+  type ProfileSavedStateTransform,
+  type TabTypeStateMap,
 } from './profile_state';
 export { createProfileStateRegistry } from './create_profile_state_registry';
 export {
@@ -26,11 +29,4 @@ export {
   METRICS_STATE_DEF,
   type MetricsState,
 } from './profile_state_definitions/metrics_grid_profile_state';
-export {
-  ProfileSavedStateRegistry,
-  createProfileSavedStateTransform,
-  type ProfileSavedStateTransform,
-  type TabTypeStateMap,
-} from './profile_saved_state';
-export { createProfileSavedStateRegistry } from './create_profile_saved_state_registry';
 export { METRICS_GRID_SAVED_STATE_TRANSFORM } from './profile_saved_state_transforms/metrics_grid_saved_state_transform';
