+++ src/platform/plugins/shared/discover/common/context_awareness/profile_state.ts
@@ -8,6 +8,8 @@
  */
 
 import type { SerializableRecord } from '@kbn/utility-types';
+import type { DiscoverTabType } from '@kbn/discover-utils';
+import type { DiscoverSessionTabTypeState } from '@kbn/saved-search-plugin/common';
 import { isEqual } from 'lodash';
 
 /**
@@ -69,6 +71,34 @@ export interface ProfileStateDefinition<TState extends SerializableRecord> {
  */
 export type ProfileStateMap = Record<string, SerializableRecord | undefined>;
 
+/** Saved payloads by tab type, derived from the saved object schema. */
+export type TabTypeStateMap = {
+  [T in DiscoverSessionTabTypeState as T['type']]: Omit<T, 'type'>;
+};
+
+/** Maps runtime profile state to and from part of a tab type's saved payload. */
+export interface ProfileSavedStateTransform<
+  TTabType extends keyof TabTypeStateMap,
+  TState extends SerializableRecord,
+  TField extends keyof TabTypeStateMap[TTabType] = keyof TabTypeStateMap[TTabType]
+> {
+  tabType: TTabType;
+  /** Fields of the tab type payload this transform owns. Must not overlap another transform's. */
+  savedFields: readonly TField[];
+  stateDefinition: ProfileStateDefinition<TState>;
+  toSavedState: (state: TState) => Pick<TabTypeStateMap[TTabType], TField>;
+  fromSavedState: (saved: Pick<TabTypeStateMap[TTabType], TField>) => Partial<TState>;
+}
+
+/** Identity helper so callers don't have to write the transform's generics by hand. */
+export const createProfileSavedStateTransform = <
+  TTabType extends keyof TabTypeStateMap,
+  TState extends SerializableRecord,
+  TField extends keyof TabTypeStateMap[TTabType]
+>(
+  transform: ProfileSavedStateTransform<TTabType, TState, TField>
+): ProfileSavedStateTransform<TTabType, TState, TField> => transform;
+
 /**
  * Controls how registered default values are handled when filtering profile state.
  */
@@ -79,6 +109,13 @@ type ProfileStateDescriptorEntry<TState extends SerializableRecord> = [
   ProfileStateDescriptor<TState>[keyof TState]
 ];
 
+interface RegisteredTransform {
+  stateDefinition: ProfileStateDefinition<SerializableRecord>;
+  savedFields: ReadonlySet<string>;
+  toSavedState: (state: SerializableRecord) => SerializableRecord;
+  fromSavedState: (saved: SerializableRecord) => Partial<SerializableRecord>;
+}
+
 const getProfileStateDescriptorEntries = <TState extends SerializableRecord>(
   descriptor: ProfileStateDescriptor<TState>
 ): Array<ProfileStateDescriptorEntry<TState>> => {
@@ -86,10 +123,11 @@ const getProfileStateDescriptorEntries = <TState extends SerializableRecord>(
 };
 
 /**
- * Registry of profile state definitions supported by Discover.
+ * Registry of profile state definitions and saved state transforms supported by Discover.
  */
 export class ProfileStateRegistry {
   private readonly stateDefinitions = new Map<string, ProfileStateDefinition<SerializableRecord>>();
+  private readonly transformsByTabType = new Map<string, RegisteredTransform[]>();
 
   /**
    * Registers a profile state definition. Keys must be globally unique.
@@ -120,6 +158,97 @@ export class ProfileStateRegistry {
       isEqual(registeredDefinition.descriptor, definition.descriptor) &&
       isEqual(registeredDefinition.defaultState, definition.defaultState)
     );
+  }
+
+  /** Registers a saved state transform for an existing matching state definition. */
+  public registerTransform<
+    TTabType extends keyof TabTypeStateMap,
+    TState extends SerializableRecord,
+    TField extends keyof TabTypeStateMap[TTabType]
+  >(transform: ProfileSavedStateTransform<TTabType, TState, TField>) {
+    if (!this.hasDefinition(transform.stateDefinition)) {
+      throw new Error(
+        `A matching state definition with key ${transform.stateDefinition.key} must be registered before its saved state transform.`
+      );
+    }
+
+    const existingTransforms = this.transformsByTabType.get(transform.tabType) ?? [];
+    const claimedFields = new Set(
+      existingTransforms.flatMap(({ savedFields }) => [...savedFields])
+    );
+
+    for (const field of transform.savedFields as readonly string[]) {
+      if (claimedFields.has(field)) {
+        throw new Error(
+          `Field "${String(field)}" of tab type "${String(
+            transform.tabType
+          )}" is already claimed by another transform.`
+        );
+      }
+    }
+
+    this.transformsByTabType.set(transform.tabType, [
+      ...existingTransforms,
+      {
+        stateDefinition: transform.stateDefinition,
+        savedFields: new Set(transform.savedFields as readonly string[]),
+        toSavedState: transform.toSavedState as unknown as (
+          state: SerializableRecord
+        ) => SerializableRecord,
+        fromSavedState: transform.fromSavedState as unknown as (
+          saved: SerializableRecord
+        ) => Partial<SerializableRecord>,
+      },
+    ]);
+  }
+
+  /** Builds saved state from each transform's effective runtime state. */
+  public toSavedState(
+    tabType: DiscoverTabType | undefined,
+    profileStateMap: ProfileStateMap
+  ): DiscoverSessionTabTypeState | undefined {
+    if (!tabType) {
+      return undefined;
+    }
+
+    const transforms = this.transformsByTabType.get(tabType) ?? [];
+    let payload: SerializableRecord = {};
+
+    for (const transform of transforms) {
+      const effectiveState = {
+        ...transform.stateDefinition.defaultState,
+        ...profileStateMap[transform.stateDefinition.key],
+      };
+      payload = { ...payload, ...transform.toSavedState(effectiveState) };
+    }
+
+    return { type: tabType, ...payload } as DiscoverSessionTabTypeState;
+  }
+
+  /** Restores registered profile state and ignores unclaimed saved fields. */
+  public fromSavedState(tabTypeState: DiscoverSessionTabTypeState | undefined): ProfileStateMap {
+    const profileStateMap: ProfileStateMap = {};
+
+    if (!tabTypeState) {
+      return profileStateMap;
+    }
+
+    const { type, ...payload } = tabTypeState as SerializableRecord & { type: string };
+    const transforms = this.transformsByTabType.get(type) ?? [];
+
+    for (const transform of transforms) {
+      const narrowedPayload: SerializableRecord = {};
+
+      for (const field of transform.savedFields) {
+        if (Object.hasOwn(payload, field)) {
+          narrowedPayload[field] = payload[field];
+        }
+      }
+
+      profileStateMap[transform.stateDefinition.key] = transform.fromSavedState(narrowedPayload);
+    }
+
+    return profileStateMap;
   }
 
   /**
