+++ src/platform/plugins/shared/discover/public/__mocks__/discover_state.mock.ts
@@ -96,7 +96,6 @@ function createInternalStateStoreMock({
     urlStateStorage: stateStorageContainer,
     storage: services.storage,
     profileStateRegistry: services.profileStateRegistry,
-    profileSavedStateRegistry: services.profileSavedStateRegistry,
   });
   const searchSessionManager = new DiscoverSearchSessionManager({
     history: services.history,
