+++ src/platform/plugins/shared/discover/public/application/main/components/top_nav/save_discover_session/save_discover_session_modal_container.test.tsx
@@ -355,7 +355,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
           },
         }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       const dataViewWithTimeFieldTab = fromTabStateToSavedObjectTab({
@@ -368,7 +368,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
           },
         }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       const adHocDataViewNoTimeFieldTab = fromTabStateToSavedObjectTab({
@@ -381,7 +381,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
           },
         }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       const adHocDataViewWithTimeFieldTab = fromTabStateToSavedObjectTab({
@@ -397,7 +397,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
           },
         }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       it("should set isTimeBased to false if no tab's data view is time based", async () => {
@@ -450,7 +450,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
         services,
         tab: getTabStateMock({ id: 'noTimeRestoreTab', attributes: { timeRestore: false } }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       const timeRestoreTab = fromTabStateToSavedObjectTab({
@@ -458,7 +458,7 @@ describe('DiscoverSessionSaveModalContainer', () => {
         services,
         tab: getTabStateMock({ id: 'timeRestoreTab', attributes: { timeRestore: true } }),
         tabType: undefined,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
+        profileStateRegistry: services.profileStateRegistry,
       });
 
       let { modalProps } = await setup({
