+++ src/platform/plugins/shared/discover/public/application/main/state_management/discover_app_state.test.ts
@@ -64,7 +64,7 @@ describe('Test discover app state', () => {
       services,
       currentDataView: undefined,
       tabType: undefined,
-      profileSavedStateRegistry: services.profileSavedStateRegistry,
+      profileStateRegistry: services.profileStateRegistry,
     });
 
     return createDiscoverSessionMock({
