+++ src/platform/plugins/shared/discover/public/application/main/state_management/hooks/use_state_managers.ts
@@ -46,7 +46,6 @@ export const useStateManagers = ({
       urlStateStorage,
       storage: services.storage,
       profileStateRegistry: services.profileStateRegistry,
-      profileSavedStateRegistry: services.profileSavedStateRegistry,
       enabled: tabsEnabled,
     })
   );
