+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/__mocks__/internal_state.mocks.ts
@@ -93,6 +93,6 @@ export const getPersistedTabMock = ({
     services,
     currentDataView: dataView,
     tabType,
-    profileSavedStateRegistry: services.profileSavedStateRegistry,
+    profileStateRegistry: services.profileStateRegistry,
   });
 };
