+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/actions/save_discover_session.test.ts
@@ -444,7 +444,7 @@ describe('saveDiscoverSession', () => {
             services,
             currentDataView: undefined,
             tabType: undefined,
-            profileSavedStateRegistry: services.profileSavedStateRegistry,
+            profileStateRegistry: services.profileStateRegistry,
           }),
         ],
       });
@@ -492,7 +492,7 @@ describe('saveDiscoverSession', () => {
             services,
             currentDataView: undefined,
             tabType: undefined,
-            profileSavedStateRegistry: services.profileSavedStateRegistry,
+            profileStateRegistry: services.profileStateRegistry,
           }),
         ],
       });
@@ -537,7 +537,7 @@ describe('saveDiscoverSession', () => {
             services,
             currentDataView: undefined,
             tabType: undefined,
-            profileSavedStateRegistry: services.profileSavedStateRegistry,
+            profileStateRegistry: services.profileStateRegistry,
           }),
         ],
       });
