+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/actions/save_discover_session.ts
@@ -75,7 +75,7 @@ export const saveDiscoverSession = createInternalStateAsyncThunk(
             currentDataView,
             services,
             tabType: selectTabTypeForPersistence({ runtimeStateManager, tabState: tab }),
-            profileSavedStateRegistry: services.profileSavedStateRegistry,
+            profileStateRegistry: services.profileStateRegistry,
           })
         );
