+++ src/platform/plugins/shared/discover/public/application/main/state_management/tabs_storage_manager.test.tsx
@@ -105,7 +105,6 @@ describe('TabsStorageManager', () => {
         urlStateStorage,
         storage: services.storage,
         profileStateRegistry: services.profileStateRegistry,
-        profileSavedStateRegistry: services.profileSavedStateRegistry,
         enabled: true,
       }),
     };
